blob-package {blob}	R Documentation

blob: A Simple S3 Class for Representing Vectors of Binary Data ('BLOBS')

Description

R's raw vector is useful for storing a single binary object. What if you want to put a vector of them in a data frame? The 'blob' package provides the blob object, a list of raw vectors, suitable for use as a column in data frame.

Author(s)

Maintainer: Kirill Müller kirill@cynkra.com

Authors:

Hadley Wickham

Other contributors:

RStudio [copyright holder, funder]

See Also

Useful links:

https://blob.tidyverse.org

https://github.com/tidyverse/blob

Report bugs at https://github.com/tidyverse/blob/issues

[Package blob version 1.3.0 Index]