Deprecated generic

The as.blob() generic has been deprecated in favor of the as_blob() function (for users) and the vec_cast.blob() method (for implementers). Implement a vec_cast.blob.myclass() method to support coercing objects of your class to blobs. See vctrs::vec_cast() for more detail.

as.blob(x, ...)

Arguments

x
An object.

...
Passed on to methods.