blob {blob}	R Documentation

Construct a blob object

Description

new_blob() is a low-level constructor that takes a list of raw vectors. blob() constructs a blob from individual raw vectors. as_blob() and is_blob() are simple forwarders to vctrs::vec_cast() and inherits(), respectively.

Usage

blob(...)

new_blob(x = list())

validate_blob(x)

as_blob(x)

is_blob(x)

Arguments

...
Individual raw vectors

x
A list of raw vectors, or other object to coerce

See Also

as.blob() for the legacy interface for specifying casts.

Examples

x1 <- charToRaw("Good morning")
x2 <- as.raw(c(0x48, 0x65, 0x6c, 0x6c, 0x6f))

new_blob(list(x1, x2))
blob(x1, x2)

as.blob(c("Good morning", "Good evening"))

[Package blob version 1.2.4 Index]